Refactor how the toolset ignores files, so that files are ignored based on semantics and not filenames; allow non-ignored files if they are specified by filename and not by directory

+++ se/commands/semanticate.py
@@ -23,19 +23,23 @@ def semanticate() -> int:
 	console = Console(highlight=False, theme=se.RICH_THEME, force_terminal=se.is_called_from_parallel()) # Syntax highlighting will do weird things when printing paths; force_terminal prints colors when called from GNU Parallel
 	return_code = 0
 
-	for filename in se.get_target_filenames(args.targets, (".xhtml",)):
+	for filename in se.get_target_filenames(args.targets, ".xhtml"):
 		if args.verbose:
 			console.print(f"Processing [path][link=file://{filename}]{filename}[/][/] ...", end="")
 
 		try:
 			with open(filename, "r+", encoding="utf-8") as file:
 				xhtml = file.read()
-				processed_xhtml = se.formatting.semanticate(xhtml)
 
-				if processed_xhtml != xhtml:
-					file.seek(0)
-					file.write(processed_xhtml)
-					file.truncate()
+				is_ignored, _ = se.get_dom_if_not_ignored(xhtml)
+
+				if not is_ignored:
+					processed_xhtml = se.formatting.semanticate(xhtml)
+
+					if processed_xhtml != xhtml:
+						file.seek(0)
+						file.write(processed_xhtml)
+						file.truncate()
 		except FileNotFoundError:
 			se.print_error(f"Couldn’t open file: [path][link=file://{filename}]{filename}[/][/].")
 			return_code = se.InvalidInputException.code
